now we can get the distance and duration from HTTP call

+++ internal-common/src/main/java/com/msb/internalcommon/response/DirectionResponse.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class DirectionResponse {
-    private Integer distance;
+    private Double distance;
     private Integer duration;
 }
